+++ Lesson08/Lesson08/Lesson08.csproj
@@ -45,6 +45,7 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ReportDatas.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
