+++ Lesson08/Lesson08/Program.cs
@@ -8,25 +8,15 @@ using System.Runtime.ExceptionServices;
 using System.Text;
 using System.Threading.Tasks;
 
+/// <summary>
+/// Алексей Кулик kpblc2000@yandex.ru
+/// Алгоритмы, Урок 8
+/// Реализовать сортировку подсчетом. Реализовать быструю сортировку.
+/// Проанализировать время работы каждого из вида сортировок для 100, 10000, 1000000 элементов.
+/// </summary>
+
 namespace Lesson08
 {
-
-	/// <summary>
-	/// Класс для хранения данных о времени выполнения.
-	/// Сделано весьма топорно - только для просмотра результатов.
-	/// </summary>
-	class ReportDatas
-	{
-
-		public int ArrayLength;
-		public double SecondsToRunQuickSort;
-		public double SecondsToRunCountSort;
-
-		public override string ToString()
-		{
-			return String.Format("{0,10}{1,20:F3}{2,20:F3}", ArrayLength, SecondsToRunQuickSort, SecondsToRunCountSort);
-		}
-	}
 
 	class Program
 	{
@@ -161,7 +151,7 @@ namespace Lesson08
 		{
 			List<ReportDatas> lst = new List<ReportDatas>();
 
-			foreach (int key in new int[] { 100, 1000, 10000, 1000000 })
+			foreach (int key in new int[] { 100, 1000, 10000, 1000000})
 			{
 				// Оставил для целей тестирования
 				int[] baseArray;
